Made credentials print statement a warning.
The "credentials were not supplied. open data access only" statement in the has_auth function used to be a print statement, which got very annoying to work with for users without credentials. Having it be a warning statement makes it easier for users to surpress these repeated statements, eg. with warnings.filterwarnings("ignore").